# perf: reduce mem allocs in scs frontend