--- a/constraint/r1cs_test.go
+++ b/constraint/r1cs_test.go
@@ -31,26 +31,26 @@ func ExampleR1CS_GetR1Cs() {
 
 	// X² == X * X
 	r1cs.AddR1C(constraint.R1C{
-		L: constraint.LinearExpression{r1cs.MakeTerm(&cOne, X)},
-		R: constraint.LinearExpression{r1cs.MakeTerm(&cOne, X)},
-		O: constraint.LinearExpression{r1cs.MakeTerm(&cOne, v0)},
+		L: constraint.LinearExpression{r1cs.MakeTerm(cOne, X)},
+		R: constraint.LinearExpression{r1cs.MakeTerm(cOne, X)},
+		O: constraint.LinearExpression{r1cs.MakeTerm(cOne, v0)},
 	}, blueprint)
 
 	// X³ == X² * X
 	r1cs.AddR1C(constraint.R1C{
-		L: constraint.LinearExpression{r1cs.MakeTerm(&cOne, v0)},
-		R: constraint.LinearExpression{r1cs.MakeTerm(&cOne, X)},
-		O: constraint.LinearExpression{r1cs.MakeTerm(&cOne, v1)},
+		L: constraint.LinearExpression{r1cs.MakeTerm(cOne, v0)},
+		R: constraint.LinearExpression{r1cs.MakeTerm(cOne, X)},
+		O: constraint.LinearExpression{r1cs.MakeTerm(cOne, v1)},
 	}, blueprint)
 
 	// Y == X³ + X + 5
 	r1cs.AddR1C(constraint.R1C{
-		R: constraint.LinearExpression{r1cs.MakeTerm(&cOne, ONE)},
-		L: constraint.LinearExpression{r1cs.MakeTerm(&cOne, Y)},
+		R: constraint.LinearExpression{r1cs.MakeTerm(cOne, ONE)},
+		L: constraint.LinearExpression{r1cs.MakeTerm(cOne, Y)},
 		O: constraint.LinearExpression{
-			r1cs.MakeTerm(&cFive, ONE),
-			r1cs.MakeTerm(&cOne, X),
-			r1cs.MakeTerm(&cOne, v1),
+			r1cs.MakeTerm(cFive, ONE),
+			r1cs.MakeTerm(cOne, X),
+			r1cs.MakeTerm(cOne, v1),
 		},
 	}, blueprint)