--- a/internal/generator/backend/template/representations/coeff.go.tmpl
+++ b/internal/generator/backend/template/representations/coeff.go.tmpl
@@ -53,9 +53,8 @@ func (ct *CoeffTable) AddCoeff(coeff constraint.Element) uint32 {
 	return cID
 }
 
-
-func (ct *CoeffTable) MakeTerm(coeff *constraint.Element, variableID int) constraint.Term {
-	cID := ct.AddCoeff(*coeff)
+func (ct *CoeffTable) MakeTerm(coeff constraint.Element, variableID int) constraint.Term {
+	cID := ct.AddCoeff(coeff)
 	return constraint.Term{VID: uint32(variableID), CID: cID}
 }
 
@@ -63,9 +62,6 @@ func (ct *CoeffTable) MakeTerm(coeff *constraint.Element, variableID int) constr
 func (ct *CoeffTable) CoeffToString(cID int) string {
 	return ct.Coefficients[cID].String()
 }
-
-
-
 
 // implements constraint.Field
 type field struct{}