--- a/internal/generator/backend/template/representations/tests/r1cs.go.tmpl
+++ b/internal/generator/backend/template/representations/tests/r1cs.go.tmpl
@@ -71,7 +71,6 @@ func TestSerialization(t *testing.T) {
 					 "field",
 					 "CoeffTable.mCoeffs",
 					 "System.lbWireLevel",
-					 "System.lbHints",
 					 "System.genericHint",
 					 "System.SymbolTable",
 					 "System.lbOutputs",